requirements.txt
et-xmlfile==1.0.1
google-api-python-client==1.5.5
httplib2==0.17.0
jdcal==1.4.1
oauth2client==4.1.3
openpyxl==3.0.3
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.0
six==1.14.0
uritemplate==3.0.1
eel
gevent
